Separated path distiction logic from DatasetType enum. Adjusted imports accordingly
